# APIdoc/api.yaml
openapi: 3.0.3
info:
  title: Documentazione API progetto "SatisTrento"
  description: Documentazione API per il progetto "SatisTrento" all'interno del progetto "100 progetti per il comune di Trento" per il corso di Ingegneria del Software a.a. 2024/2025 corso di laurea in Informatica presso l'Università degli Studi di Trento.
  version: 0.0.1
servers:
  - url: http://localhost:3000/api/v1
    description: Development server for testing
components:
  schemas:
    DatiBase:
      description: Oggetto contenente le informazioni di base di un quartiere.
      type: object
      properties:
        popolazione:
            type: integer
            description: Numero di abitanti del quartiere.
        superficie:
          type: number
          description: Superficie del quartiere in km^2.
          format: double
        serviziTotali: 
          type: integer
          description: Numero totale di servizi presenti nel quartiere.
        soddisfazioneMedia:
          type: number
          description: Soddisfazione media degli abitanti del quartiere.
          format: double
          minimum: 0
          maximum: 100
        interventiPolizia: 
          type: integer
          description: Numero di interventi delle forze dell'ordine nel quartiere.
        etaMedia:
          type: number
          description: Età media degli abitanti del quartiere.
          format: double
    ServiziGenerali:
      description: Oggetto contenente le informazioni sui servizi presenti in un quartiere.
      type: object
      properties:
        areeVerdi:
          type: integer
          description: Numero di aree verdi presenti nel quartiere.
          minimum: 0
        scuole:
          type: integer
          description: Numero di scuole presenti nel quartiere.
          minimum: 0
        serviziRistorazione:
          type: integer
          description: Numero di servizi di ristorazione presenti nel quartiere.
          minimum: 0
        localiNotturni: 
          type: integer
          description: Numero di locali notturni presenti nel quartiere.
          minimum: 0
    Sicurezza:
      description: Oggetto contenente le informazioni sulla sicurezza di un quartiere.
      type: object
      properties:
        numeroInterventi:
          type: integer
          description: Numero di interventi delle forze dell'ordine nel quartiere.
          minimum: 0
        incidenti: 
          type: integer
          description: Numero di incidenti nel quartiere.
          minimum: 0
        tassoCriminalita:
          type: number
          description: Tasso di criminalità del quartiere.
          format: double
          minimum: 0
    Coordinate:
      description: Array di coordinate che definiscono un quartiere e/o circoscrizione.
      type: array
      items: 
        description: Coordinate nel formato [latitude, longitude].
        type: array
        items:
          type: number
          format: double
        minItems: 2
        maxItems: 2
      minItems: 1
      uniqueItems: true
    Soddisfazione: 
        type: "number"
        description: "Soddisfazione media degli abitanti del quartiere."
        format: "double"
        minimum: 0
        maximum: 100
    CircoscrizioneBase: 
      description: Oggetto contenente le informazioni di Base di una circoscrizione.
      type: object
      properties:
        _id:
          type: string
          description: Identificatore univoco della circoscrizione.
        nome:
          type: string
          description: Nome della circoscrizione.
        coordinate:
          "$ref": "#/components/schemas/Coordinate"
        soddisfazioneMedia:
          "$ref": "#/components/schemas/Soddisfazione"
...
    QuartiereBase: 
      description: Oggetto contenente le informazioni di base di un quartiere.
      type: object
      properties:
        _id:
          type: string
          description: "Identificatore univoco del quartiere."
        nome:
          type: "string"
          description: "Nome del quartiere."
        coordinate:
          "$ref": "#/components/schemas/Coordinate"
        circoscrizione:
          "$ref": "#/components/schemas/CircoscrizioneBase"
        soddisfazioneMedia:
          "$ref": "#/components/schemas/Soddisfazione"
    Quartiere: 
      description: Oggetto contenente le informazioni di un quartiere.
      allOf:
        - "$ref": "#/components/schemas/QuartiereBase"
        - "$ref": "#/components/schemas/DatiBase"
      type: object
      properties:
        servizi:
          "$ref": "#/components/schemas/ServiziGenerali"
        sicurezza:
          "$ref": "#/components/schemas/Sicurezza"
...
    Error:
      description: Oggetto contenente le informazioni di un errore.
      type: object
      properties:
        code:
          type: integer
          description: Codice dell'errore.
        message:
          type: string
          description: Messaggio di errore.
        details:
          type: string
          description: Dettagli dell'errore.
          nullable: true
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token di autenticazione fornito all'utente per accedere alle risorse specifiche protette.
paths:
  '/quartieri' :
    description: "Risorsa che restituisce la lista dei quartieri di Trento."
    get: 
      summary: "Ottieni lista dei quartieri"
      description: "Ritorna una lista di tutti i quartieri della città di Trento."
      parameters:
        - name: deepData
          in: query
          description: "Se impostato a true, restituisce le informazioni complete di ogni quartiere."
          required: false
          schema:
            type: boolean
            default: false
          allowEmptyValue: true
      responses: 
        200: 
          description: "Risposta con la lista dei quartieri."
          content:
            application/json:
              schema: 
                oneOf:
                  - type: array
                    items: 
                      "$ref": "#/components/schemas/QuartiereBase"
                  - type: array
                    items: 
                      "$ref": "#/components/schemas/Quartiere"
      tags:
        - Quartieri
        - Dati Statici
  '/quartieri/{id}' :
    description: "Risorsa che restituisce le informazioni di un quartiere di Trento con le informazioni di base e i servizi presenti."
    parameters:
      - name: id
        in: path
        description: "Identificatore univoco del quartiere."
        required: true
        schema:
          type: string
    get: 
      summary: "Ottieni informazioni di un quartiere"
      responses: 
        200:
          description: "Risposta con le informazioni del quartiere."
          content:
            application/json: 
              schema: 
                $ref: "#/components/schemas/Quartiere"
        404: 
          description: "Quartiere non trovato."
          content: 
            application/json: 
              schema: 
                "$ref": "#/components/schemas/Error"
      tags:
        - Quartieri
        - Dati Statici
...